+++ src/components/Navigation.vue
@@ -43,83 +43,16 @@
 </template>
 
 <script lang="ts">
+import { Modal } from "../modal";
+
 export default {
 	props: {
 		listOfTripNames: { required: true, type: Array }
 	},
 	methods: {
-		// TODO: should we rather have the root Vue instance handle creating a modal? 
-		// then we can just call a method from root and pass it params to customize as needed?
-		showPrompt: async function(): Promise<string> {
-			const modal = document.createElement('div');
-			modal.id = "modal";
-			modal.style.position = "fixed";
-			modal.style.top = "50%";
-			modal.style.left = "50%";
-			modal.style.transform = "translate(-50%, -50%)";
-			modal.style.zIndex = "1010";
-			modal.style.textAlign = "center";
-			modal.style.padding = "5px";
-			modal.style.backgroundColor = "#fff";
-			modal.style.width = "25%";
-			modal.style.height = "25%";
-			modal.style.boxShadow = "2px 2px 5px #ccc";
-			
-			const text = document.createElement('h2');
-			const text2 = document.createElement('p');
-			
-			text.textContent = "this is a test";
-			text2.textContent = "please enter the name of the new trip:";
-			
-			const textInput = document.createElement('input');
-			textInput.type = "text";
-			
-			modal.appendChild(text);
-			modal.appendChild(text2);
-			modal.appendChild(textInput);
-			
-			const submitBtn = document.createElement('button');
-			submitBtn.innerText = "submit";
-			
-			const cancelBtn = document.createElement('button');
-			cancelBtn.innerText = "cancel";
-			
-			modal.appendChild(submitBtn);
-			modal.appendChild(cancelBtn);
-			
-			const modalOverlay = document.createElement('div');
-			modalOverlay.id = "modal-overlay";
-			modalOverlay.style.zIndex = "1000";
-			modalOverlay.style.position = "fixed";
-			modalOverlay.style.top = "0";
-			modalOverlay.style.left = "0";
-			modalOverlay.style.width = "100%";
-			modalOverlay.style.height = "100%";
-			modalOverlay.style.backgroundColor = "#aaa";
-			modalOverlay.style.opacity = "0.2";
-			
-			document.body.appendChild(modal);
-			document.body.appendChild(modalOverlay);
-			
-			return new Promise<string>((resolve, reject) => {
-				submitBtn.onclick = () => {
-					// get the input text and resolve it
-					resolve("this is a test");
-				};
-				
-				cancelBtn.onclick = () => {
-					resolve(undefined);
-				};
-			}).finally(() => {
-				// make sure to close modal
-				document.body.removeChild(modal);
-				document.body.removeChild(modalOverlay);
-			});
-		},
-	
 		addNewTrip: async function(): Promise<void> {
-			const newTripName = await this.showPrompt();
-			//console.log(newTripName);
+			const modalHandler = new Modal();
+			const newTripName = await modalHandler.createInputModal("please enter the name of the new trip:");
 			
 			if(newTripName) {
 				//@ts-ignore (TS-2339)
